requirements.txt
asyncio
websockets
protobuf==3.20.*
google
protobuf-decoder
pyyaml
pydantic
pytest